moved and refactored code to better organise files
