+++ .gitignore
@@ -1,2 +1,4 @@
 *.pyc
 .github/.DS_Store
+__pycache__
+.ipynb_checkpoints
