--- UI/metadata_cache_Albertsons.json
@@ -1,1 +1,0 @@
-{"camain_oracle_hcm": {"table_name": "camain_oracle_hcm", "columns_info": {"fields": [{"name": "EFFECTIVE_START_DATE", "type": "DATE"}, {"name": "EFFECTIVE_END_DATE", "type": "DATE"}, {"name": "COLLECTIVE_AGREEMENT_NAME", "type": "STRING"}, {"name": "IDENTIFICATION_CODE", "type": "STRING"}, {"name": "COMMENTS", "type": "STRING"}, {"name": "DESCRIPTION", "type": "STRING"}, {"name": "UNION_NAME", "type": "STRING"}, {"name": "BARGAINING_UNIT_DESCRIPTION", "type": "STRING"}, {"name": "STATUS", "type": "STRING"}, {"name": "AI_EXTRACTION_TIMESTAMP", "type": "TIMESTAMP"}, {"name": "SPECIAL_LABOR_AGREEMENT", "type": "BOOLEAN"}, {"name": "RATIFICATION_ID", "type": "STRING"}, {"name": "PRIMARY_LABOR_AGREEMENT", "type": "STRING"}, {"name": "RATIFICATION_DATE", "type": "TIMESTAMP"}, {"name": "CONTRACT_BEGIN_DATE", "type": "TIMESTAMP"}, {"name": "CONTRACT_END_DATE", "type": "TIMESTAMP"}]}, "table_description": "The table `camain_oracle_hcm` appears to store information about collective labor agreements, including their effective dates, names, descriptions, union details, ratification status, and contract durations. \n"}, "camain_ps": {"table_name": "camain_ps", "columns_info": {"fields": [{"name": "AI_EXTRACTION_TIMESTAMP", "type": "TIMESTAMP"}, {"name": "RATIFICATION_ID", "type": "STRING"}, {"name": "SET_ID", "type": "STRING"}, {"name": "LABOR_AGREEMENT", "type": "STRING"}, {"name": "EFFECTIVE_DATE", "type": "DATE"}, {"name": "STATUS_AS_OF_EFFECTIVE_DATE", "type": "STRING"}, {"name": "UNION_CODE", "type": "INTEGER"}, {"name": "BARGAINING_UNIT", "type": "INTEGER"}, {"name": "DESCRIPTION", "type": "STRING"}, {"name": "SHORT_DESCRIPTION", "type": "STRING"}, {"name": "WAGE_PROGRESSION_HOURS_WORKED", "type": "BOOLEAN"}, {"name": "STEP_ID", "type": "BOOLEAN"}, {"name": "CARRYOVR_WP_STP_DT", "type": "BOOLEAN"}, {"name": "SPECIAL_LABOR_AGREEMENT", "type": "BOOLEAN"}, {"name": "COMMENTS", "type": "STRING"}, {"name": "LBR_EXPIRY_RULE_ID", "type": "STRING"}, {"name": "SENIORITY_RULE_FLG", "type": "BOOLEAN"}, {"name": "FACILTY_FLG", "type": "BOOLEAN"}, {"name": "CONTRACT_RATIFIED_DATE", "type": "DATE"}, {"name": "CONTRACT_BEGIN_DATE", "type": "DATE"}, {"name": "CONTRACT_END_DATE", "type": "DATE"}]}, "table_description": "The table `camain_ps` appears to store information about labor agreements, including their effective dates, ratification status, associated unions and bargaining units, descriptions, and various flags indicating specific clauses or attributes. \n"}}
