--- UI/metadata_cache_thelook_ecommerce.json
@@ -1,1 +1,0 @@
-{"distribution_centers": {"table_name": "distribution_centers", "columns_info": {"fields": [{"name": "id", "type": "INTEGER"}, {"name": "name", "type": "STRING"}, {"name": "latitude", "type": "FLOAT"}, {"name": "longitude", "type": "FLOAT"}]}, "table_description": "The table \"distribution_centers\" stores information about distribution centers, including their unique ID, name, and geographical location (latitude and longitude). \n"}, "events": {"table_name": "events", "columns_info": {"fields": [{"name": "id", "type": "INTEGER"}, {"name": "user_id", "type": "INTEGER"}, {"name": "sequence_number", "type": "INTEGER"}, {"name": "session_id", "type": "STRING"}, {"name": "created_at", "type": "TIMESTAMP"}, {"name": "ip_address", "type": "STRING"}, {"name": "city", "type": "STRING"}, {"name": "state", "type": "STRING"}, {"name": "postal_code", "type": "STRING"}, {"name": "browser", "type": "STRING"}, {"name": "traffic_source", "type": "STRING"}, {"name": "uri", "type": "STRING"}, {"name": "event_type", "type": "STRING"}]}, "table_description": "The `events` table stores user activity data, including session details, timestamps, location information, browser used, traffic source, and the specific type of event triggered. \n"}, "inventory_items": {"table_name": "inventory_items", "columns_info": {"fields": [{"name": "id", "type": "INTEGER"}, {"name": "product_id", "type": "INTEGER"}, {"name": "created_at", "type": "TIMESTAMP"}, {"name": "sold_at", "type": "TIMESTAMP"}, {"name": "cost", "type": "FLOAT"}, {"name": "product_category", "type": "STRING"}, {"name": "product_name", "type": "STRING"}, {"name": "product_brand", "type": "STRING"}, {"name": "product_retail_price", "type": "FLOAT"}, {"name": "product_department", "type": "STRING"}, {"name": "product_sku", "type": "STRING"}, {"name": "product_distribution_center_id", "type": "INTEGER"}]}, "table_description": "The `inventory_items` table tracks individual items in inventory, recording their product details, cost, purchase and sale dates, and associated distribution center. \n"}, "order_items": {"table_name": "order_items", "columns_info": {"fields": [{"name": "id", "type": "INTEGER"}, {"name": "order_id", "type": "INTEGER"}, {"name": "user_id", "type": "INTEGER"}, {"name": "product_id", "type": "INTEGER"}, {"name": "inventory_item_id", "type": "INTEGER"}, {"name": "status", "type": "STRING"}, {"name": "created_at", "type": "TIMESTAMP"}, {"name": "shipped_at", "type": "TIMESTAMP"}, {"name": "delivered_at", "type": "TIMESTAMP"}, {"name": "returned_at", "type": "TIMESTAMP"}, {"name": "sale_price", "type": "FLOAT"}]}, "table_description": "The `order_items` table stores information about individual items within an order, including the associated order, user, and product IDs, inventory details, order status, timestamps for various stages of the order fulfillment process, and the sale price. \n"}, "orders": {"table_name": "orders", "columns_info": {"fields": [{"name": "order_id", "type": "INTEGER"}, {"name": "user_id", "type": "INTEGER"}, {"name": "status", "type": "STRING"}, {"name": "gender", "type": "STRING"}, {"name": "created_at", "type": "TIMESTAMP"}, {"name": "returned_at", "type": "TIMESTAMP"}, {"name": "shipped_at", "type": "TIMESTAMP"}, {"name": "delivered_at", "type": "TIMESTAMP"}, {"name": "num_of_item", "type": "INTEGER"}]}, "table_description": "The `orders` table stores information about customer orders, including order status, customer gender, timestamps for order creation, return, shipping, and delivery, as well as the number of items in each order. \n"}, "products": {"table_name": "products", "columns_info": {"fields": [{"name": "id", "type": "INTEGER"}, {"name": "cost", "type": "FLOAT"}, {"name": "category", "type": "STRING"}, {"name": "name", "type": "STRING"}, {"name": "brand", "type": "STRING"}, {"name": "retail_price", "type": "FLOAT"}, {"name": "department", "type": "STRING"}, {"name": "sku", "type": "STRING"}, {"name": "distribution_center_id", "type": "INTEGER"}]}, "table_description": "The table \"products\" stores information about products, including their cost, category, name, brand, retail price, department, SKU, and distribution center ID. \n"}, "users": {"table_name": "users", "columns_info": {"fields": [{"name": "id", "type": "INTEGER"}, {"name": "first_name", "type": "STRING"}, {"name": "last_name", "type": "STRING"}, {"name": "email", "type": "STRING"}, {"name": "age", "type": "INTEGER"}, {"name": "gender", "type": "STRING"}, {"name": "state", "type": "STRING"}, {"name": "street_address", "type": "STRING"}, {"name": "postal_code", "type": "STRING"}, {"name": "city", "type": "STRING"}, {"name": "country", "type": "STRING"}, {"name": "latitude", "type": "FLOAT"}, {"name": "longitude", "type": "FLOAT"}, {"name": "traffic_source", "type": "STRING"}, {"name": "created_at", "type": "TIMESTAMP"}]}, "table_description": "The `users` table stores information about users, including their personal details (name, age, gender, contact information), location (address, coordinates), traffic source, and creation timestamp. \n"}}
